Fixed the test weirdness.
Since test cases are performed outside the context of a TurboGears
HTTP request (and therefore without the TG app stack in place), we
must perform manual transaction management via the transaction
module. Within the TG app stack this is handled by transaction-
manager middleware, but we don't have that available within our
test cases. What this means is that whenever you *modify the
DBSession* (eg by adding a DB entity), you must wrap that in

try:
   thing = model.Whatever(....)
   model.DBSession.add(thing)
except:
   transaction.abort()
else:
   transaction.commit()

We can get TG involved (for methods within a TestController) by doing
self.app.get("/some_url"), in which case that request will be
processed as an HTTP transaction. That's great for testing
accessible URLs within a controller, but not so great for
piecemeal unit testing.

test_need_events.py is more of a unit test, but there does not
seem to be a huge advantage at the moment to establishing a
whole new subpackage for unit testing and then having to
manually set everything up in the DB. Also that seems to conflict
with TG's testing framework in ways that I don't understand
completely. So I'm going to add new tests under tests/functional
and just be sure to do the transaction stuff when necessary.