Implement volunteer activation. Test add_volunteer_start.
This commit enforces that newly-added volunteers are not eligible
to respond to events until approved by a coordinator, unless
the user account was created by a coordinator, in which case it
starts pre-activated but can be de-activated.

It also adds test cases for various use cases of the
/add_volunteer_start page, which is used for both adding and
editing volunteer information.