Hacky work-around for i18n issues.
The _ and l_ translation functions provided by TG only work
inside web requests. Since we have many tests that invoke
our utility functions outside web requests, those functions
break in the test suite. So I have written hacked versions
that check for the breakage and return the untranslated string
in broken cases.

This is bad because it incurs a silly overhead for every
translation call, but I'm presently at sea about what is the
"right" way to solve this.